Added request body parsing for json and query strings
POST, PUT, and OPTIONS requests can all specify request bodies, and the
format of those bodies is determined by the Content-Type request header.
The body parsing middleware now parses the request body string based on
the provided Content-Type, and if the result is an array, the values are
unioned with the existing request parameters.