Added content negotiation for response bodies
The request's Accept header is parsed based on the HTTP 1.1 spec into an array
of media ranges that are acceptable.  The response object takes in a callable
body parser that is used to parse the response data into a string response with
the appropriate Content-Type. Headers are compiled into their raw forms and the
body parser is actually invoked when the build() method is called on the
response. The Strapi runtime is now responsible for sending the raw headers and
outputting the [parsed] response body.

The implementation is a little rough and could likely benefit from some dev
iterations. Once an event system is in place, the body parsing can likely be
tied directly into that.