Moved response handling and content negotiation to middleware
By moving the response handling out of the strapi runtime, the existing
middleware system can be leveraged to perform custom functionality after a
request has already been dispatched to its appropriate route callback. This
made it fairly trivial to move the existing content-negotiation functionality
out of a "body parser" callback that had to be passed directly to the response
and into an optional middleware component.

As a result, the response no longer needs to be "built" prior to being
outputted, so it more accurately represents the current response state rather
following more of a builder pattern.