Added exception handling via middleware
Prior to this, any exception thrown would go uncaught and would thus trigger
a fatal error. The included exception handling middleware wraps the rest of the
middleware chain in a try/catch block and sends an appropriate response and
status code.

In addition to handling generic exceptions, a new HttpErrorException can be
thrown that is mapped directly to the HTTP response.  With this exception, the
code is mapped to the response status code and the message is mapped to the
response body.

The error handler middleware optionally will render out exception details based
on configuration directives.